requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.6
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.27.0
pywebview>=5.0
qrcode[pil]>=7.4
PyJWT>=2.8.0
bcrypt>=4.0.0
